python-service/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
python-dotenv==1.2.2
supadata==1.6.0
pinecone==8.1.0
langchain-core==1.2.23
langchain-nvidia-ai-endpoints==1.2.1
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.1
youtube-transcript-api==1.2.4